LESSON TWO: Think Like a Tester
In this lesson, we disrupt our natural instinct, to go straight to the
text editor and immediately write code to solve a coding challenge.
Instead, we follow a problem solving methodology in order to identify
the scope of the problem, inputs, outputs, examples, pseudocode,
gothcas, and ways we might test these aspects of the problem to confirm
we are deriving the right solution.