LESSON TWO: Think Like a Tester SOLUTION, PART II
In the second step of our solution, we describe the expectations of our
tests. The spec in this example is complete, but the expectations are
commented out. We will walk through a test-by-test example of
red-green-refactor by removing the comment from one test at a time,
making that test pass, then moving on to the next.