LESSON THREE: Making it Work, Step 01
In this lesson, we write the necessary code to make our first test pass
and uncomment the second one. Notice how our empty descriptions in our
test file appear to be passing tests. For this reason, you should avoid
writing them, or comment them out completely, until you fill them with
expectations. We will do that in the next step.